expr: use a 'long' for 'Exnode_t.op'
The operation constants are small numbers. But Flex generates these as `long`
values. Aligning with this squashes 8 -Wconversion warnings. This required
adjusting the types in some other code that interacts with `Exnode_t.op` too.

+++ lib/expr/exgram.h
@@ -50,7 +50,7 @@ static int T(long t) {
  * allocate and initialize a new expression node in the current program
  */
 
-Exnode_t *exnewnode(Expr_t *p, int op, bool binary, long type, Exnode_t *left,
+Exnode_t *exnewnode(Expr_t *p, long op, bool binary, long type, Exnode_t *left,
                     Exnode_t *right) {
 	Exnode_t*	x;
